Move structured field into its own module under tmt.utils
Another round of tmt.utils liberations.
