+++ pyproject.toml
@@ -252,6 +252,7 @@ ignore = [
     "tmt/lint.py",
     "tmt/queue.py",
     "tmt/utils/__init__.py",
+    "tmt/utils/structured_field.py",
     "tmt/hardware.py",  # pyright does not pick up pint's _typing.py or something :/
 ]
